feat: support complex predicate search for shared IVF and profiling complex predicate search for Curator
This commit adds comprehensive complex predicate filtering capabilities to FAISS multitenant indexes:

## MultiTenantIndexIVFFlat enhancements:
- Added search() method with complex predicate filter parameter
- Added search_preassigned() method for pre-quantized searches with predicates
- Integrated ComplexPredicateSelector for predicate evaluation during search
- Support for combining complex predicates with existing IDSelector filters
- Full parallel search support with proper error handling

## MultiTenantIndexIVFHierarchical enhancements:
- Added detailed profiling support for search_with_bitmap_filter operations
- Profiling tracks: preprocessing, sorting, temp index building, and search phases
- Added SearchProfilingData structure for comprehensive timing analysis
- Enhanced API with set_profiling_enabled() and get_last_search_profile() methods
- Improved performance monitoring for bitmap-based filtering

## Python wrapper improvements:
- Fixed dtype consistency issue in class_wrappers.py (int64 -> uint32)
- Ensures proper type compatibility across Python-C++ interface

+++ 3rd_party/faiss/faiss/python/class_wrappers.py
@@ -1308,7 +1308,7 @@ def handle_MultiTenantIndex(the_class):
         filter_key : str
             String identifier for this filter for caching purposes.
         """
-        qualified_labels = np.ascontiguousarray(qualified_labels, dtype="int64")
+        qualified_labels = np.ascontiguousarray(qualified_labels, dtype="uint32")
         n_labels = qualified_labels.shape[0]
 
         self.build_index_for_filter_c(swig_ptr(qualified_labels), n_labels, filter_key)
